Use named type for generating server or client
Separating the functions makes more semantic sense given that, as
currently used, all callers use a hard-coded constant when calling
the function. If the caller always knows the "mode" of the function
being called at compile-time, it's good signal that these should just
be separate functions (which also reduces number of parameters and
makes the behavior clearer based on the name).

Only refactors internal private functions.

==NO_CHANGELOG==